Refactor spanfilter into its own package
