+++ modules/generator/processor/spanmetrics/spanmetrics.go
@@ -1,10 +1,7 @@
-pt st
-ackage spanmetrics
+package spanmetrics
 
 import (
 	"context"
-	"reflect"
-	"regexp"
 	"time"
 
 	"github.com/opentracing/opentracing-go"
@@ -13,12 +10,10 @@ import (
 	gen "github.com/grafana/tempo/modules/generator/processor"
 	processor_util "github.com/grafana/tempo/modules/generator/processor/util"
 	"github.com/grafana/tempo/modules/generator/registry"
-	"github.com/grafana/tempo/pkg/sharedconfig"
+	"github.com/grafana/tempo/pkg/spanfilter"
 	"github.com/grafana/tempo/pkg/tempopb"
-	v1_common "github.com/grafana/tempo/pkg/tempopb/common/v1"
 	v1 "github.com/grafana/tempo/pkg/tempopb/resource/v1"
 	v1_trace "github.com/grafana/tempo/pkg/tempopb/trace/v1"
-	"github.com/grafana/tempo/pkg/traceql"
 	tempo_util "github.com/grafana/tempo/pkg/util"
 )
 
@@ -36,25 +31,13 @@ type Processor struct {
 
 	spanMetricsCallsTotal       registry.Counter
 	spanMetricsDurationSeconds  registry.Histogram
-	spanMetricsSizeTotal        registry.Countetr
+	spanMetricsSizeTotal        registry.Counter
 	spanMetricsFilterDropsTotal registry.Counter
 
-	filterPolicies []*filterPolicy
+	filter *spanfilter.SpanFilter
 
 	// for testing
 	now func() time.Time
-}
-
-type filterPolicy struct {
-	Include *splitPolicy
-	Exclude *splitPolicy
-}
-
-// SplitPolicy is the result of parsing a policy from the config file to be specific about the area the given policy is applied to.
-type splitPolicy struct {
-	ResourceMatch  *sharedconfig.PolicyMatch
-	SpanMatch      *sharedconfig.PolicyMatch
-	IntrinsicMatch *sharedconfig.PolicyMatch
 }
 
 func New(cfg Config, registry registry.Registry) gen.Processor {
@@ -91,23 +74,16 @@ func New(cfg Config, registry registry.Registry) gen.Processor {
 		p.spanMetricsSizeTotal = registry.NewCounter(metricSizeTotal, labels)
 	}
 
-	var filterPolicies []*filterPolicy
-
-	for _, policy := range cfg.FilterPolicies {
-		p := &filterPolicy{
-			Include: getSplitPolicy(policy.Include),
-			Exclude: getSplitPolicy(policy.Exclude),
-		}
-
-		if p.Include != nil || p.Exclude != nil {
-			filterPolicies = append(filterPolicies, p)
-		}
+	filter, err := spanfilter.NewSpanFilter(cfg.FilterPolicies)
+	if err != nil {
+		// TODO: handle this better
+		panic(err)
 	}
 
 	p.Cfg = cfg
 	p.registry = registry
 	p.now = time.Now
-	p.filterPolicies = filterPolicies
+	p.filter = filter
 	return p
 }
 
@@ -132,11 +108,11 @@ func (p *Processor) aggregateMetrics(resourceSpans []*v1_trace.ResourceSpans) {
 
 		for _, ils := range rs.ScopeSpans {
 			for _, span := range ils.Spans {
-				if p.applyFilterPolicy(rs.Resource, span) {
+				if p.filter.ApplyFilterPolicy(rs.Resource, span) {
 					p.aggregateMetricsForSpan(svcName, rs.Resource, span)
-				} else {
-					p.spanMetricsFilterDropsTotal.Inc(nil, 1)
+					continue
 				}
+				p.spanMetricsFilterDropsTotal.Inc(nil, 1)
 			}
 		}
 	}
@@ -185,118 +161,6 @@ func (p *Processor) aggregateMetricsForSpan(svcName string, rs *v1.Resource, spa
 	}
 }
 
-// applyFilterPolicy returns true if the span should be included in the metrics.
-func (p *Processor) applyFilterPolicy(rs *v1.Resource, span *v1_trace.Span) bool {
-	// With no filter policies specified, all spans are included.
-	if len(p.Cfg.FilterPolicies) == 0 {
-		return true
-	}
-
-	for _, policy := range p.filterPolicies {
-		if policy.Include != nil {
-			if !policyMatch(policy.Include, rs, span) {
-				return false
-			}
-		}
-
-		if policy.Exclude != nil {
-			if policyMatch(policy.Exclude, rs, span) {
-				return false
-			}
-		}
-	}
-
-	return true
-}
-
-// policyMatch returns true when the resource attribtues and span attributes match the policy.
-func policyMatch(policy *splitPolicy, rs *v1.Resource, span *v1_trace.Span) bool {
-	return policyMatchAttrs(policy.ResourceMatch, rs.Attributes) &&
-		policyMatchAttrs(policy.SpanMatch, span.Attributes) &&
-		policyMatchIntrinsicAttrs(policy.IntrinsicMatch, span)
-}
-
-// policyMatchIntrinsicAttrs returns true when all intrinsic values in the policy match the span.
-func policyMatchIntrinsicAttrs(policy *sharedconfig.PolicyMatch, span *v1_trace.Span) bool {
-	matches := 0
-	for _, pa := range policy.Attributes {
-		attr := traceql.MustParseIdentifier(pa.Key)
-		switch attr.Intrinsic {
-		// case traceql.IntrinsicDuration:
-		// case traceql.IntrinsicChildCount:
-		// case traceql.IntrinsicParent:
-		case traceql.IntrinsicName:
-			if stringMatch(policy.MatchType, span.GetName(), pa.Value.(string)) {
-				matches++
-			}
-		case traceql.IntrinsicStatus:
-			if stringMatch(policy.MatchType, span.GetStatus().GetCode().String(), pa.Value.(string)) {
-				matches++
-			}
-		case traceql.IntrinsicKind:
-			if stringMatch(policy.MatchType, span.GetKind().String(), pa.Value.(string)) {
-				matches++
-			}
-		}
-	}
-
-	return len(policy.Attributes) == matches
-}
-
-// policyMatchAttrs returns true if all attributes in the policy match the attributes in the span.  String, bool, int, and floats are supported.  Regex MatchType may be applied to string span attributes.
-func policyMatchAttrs(policy *sharedconfig.PolicyMatch, attrs []*v1_common.KeyValue) bool {
-
-	matches := 0
-	var v *v1_common.AnyValue
-
-	for _, pa := range policy.Attributes {
-		pAttrValueType := reflect.TypeOf(pa.Value).String()
-
-		for _, attr := range attrs {
-			if attr.GetKey() == pa.Key {
-				v = attr.GetValue()
-
-				switch v.Value.(type) {
-				case *v1_common.AnyValue_StringValue:
-					if pAttrValueType != "string" {
-						continue
-					}
-
-					if stringMatch(policy.MatchType, v.GetStringValue(), pa.Value.(string)) {
-						matches++
-					}
-				case *v1_common.AnyValue_IntValue:
-					if pAttrValueType != "int" {
-						continue
-					}
-
-					if v.GetIntValue() == int64(pa.Value.(int)) {
-						matches++
-					}
-				case *v1_common.AnyValue_DoubleValue:
-					if pAttrValueType != "float64" {
-						continue
-					}
-
-					if v.GetDoubleValue() == pa.Value.(float64) {
-						matches++
-					}
-				case *v1_common.AnyValue_BoolValue:
-					if pAttrValueType != "bool" {
-						continue
-					}
-
-					if v.GetBoolValue() == pa.Value.(bool) {
-						matches++
-					}
-				}
-			}
-		}
-	}
-
-	return len(policy.Attributes) == matches
-}
-
 func sanitizeLabelNameWithCollisions(name string) string {
 	sanitized := strutil.SanitizeLabelName(name)
 
@@ -313,65 +177,4 @@ func isIntrinsicDimension(name string) bool {
 		name == dimSpanKind ||
 		name == dimStatusCode ||
 		name == dimStatusMessage
-}
-
-func stringMatch(matchType sharedconfig.MatchType, s, pattern string) bool {
-	switch matchType {
-	case sharedconfig.Strict:
-		return s == pattern
-	case sharedconfig.Regex:
-		re := regexp.MustCompile(pattern)
-		return re.MatchString(s)
-	default:
-		return false
-	}
-}
-
-func getSplitPolicy(policy *sharedconfig.PolicyMatch) *splitPolicy {
-	if policy == nil {
-		return nil
-	}
-
-	// A policy to match against the resource attributes
-	resourcePolicy := &sharedconfig.PolicyMatch{
-		MatchType:  policy.MatchType,
-		Attributes: make([]sharedconfig.MatchPolicyAttribute, 0),
-	}
-
-	// A policy to match against the span attributes
-	spanPolicy := &sharedconfig.PolicyMatch{
-		MatchType:  policy.MatchType,
-		Attributes: make([]sharedconfig.MatchPolicyAttribute, 0),
-	}
-
-	intrinsicPolicy := &sharedconfig.PolicyMatch{
-		MatchType:  policy.MatchType,
-		Attributes: make([]sharedconfig.MatchPolicyAttribute, 0),
-	}
-
-	for _, pa := range policy.Attributes {
-		attr := traceql.MustParseIdentifier(pa.Key)
-
-		attribute := sharedconfig.MatchPolicyAttribute{
-			Key:   attr.Name,
-			Value: pa.Value,
-		}
-
-		if attr.Intrinsic > 0 {
-			intrinsicPolicy.Attributes = append(intrinsicPolicy.Attributes, attribute)
-		} else {
-			switch attr.Scope {
-			case traceql.AttributeScopeSpan:
-				spanPolicy.Attributes = append(spanPolicy.Attributes, attribute)
-			case traceql.AttributeScopeResource:
-				resourcePolicy.Attributes = append(resourcePolicy.Attributes, attribute)
-			}
-		}
-	}
-
-	return &splitPolicy{
-		ResourceMatch:  resourcePolicy,
-		SpanMatch:      spanPolicy,
-		IntrinsicMatch: intrinsicPolicy,
-	}
 }
