+++ modules/generator/processor/spanmetrics/spanmetrics_test.go
@@ -13,10 +13,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/tempo/modules/generator/registry"
-	"github.com/grafana/tempo/pkg/sharedconfig"
+	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 	"github.com/grafana/tempo/pkg/tempopb"
 	common_v1 "github.com/grafana/tempo/pkg/tempopb/common/v1"
-	v1 "github.com/grafana/tempo/pkg/tempopb/resource/v1"
 	trace_v1 "github.com/grafana/tempo/pkg/tempopb/trace/v1"
 	"github.com/grafana/tempo/pkg/util/test"
 )
@@ -158,19 +157,19 @@ func TestSpanMetrics_collisions(t *testing.T) {
 
 func TestSpanMetrics_applyFilterPolicy(t *testing.T) {
 	cases := []struct {
-		filterPolicies     []sharedconfig.FilterPolicy
+		filterPolicies     []filterconfig.FilterPolicy
 		expectedMatches    float64
 		expectedRejections float64
 	}{
 		{
 			expectedMatches:    10.0,
 			expectedRejections: 0.0,
-			filterPolicies: []sharedconfig.FilterPolicy{
+			filterPolicies: []filterconfig.FilterPolicy{
 				{
 
-					Include: &sharedconfig.PolicyMatch{
-						MatchType: sharedconfig.Strict,
-						Attributes: []sharedconfig.MatchPolicyAttribute{
+					Include: &filterconfig.PolicyMatch{
+						MatchType: filterconfig.Strict,
+						Attributes: []filterconfig.MatchPolicyAttribute{
 							{
 								Key:   "span.foo",
 								Value: "foo-value",
@@ -183,12 +182,12 @@ func TestSpanMetrics_applyFilterPolicy(t *testing.T) {
 		{
 			expectedMatches:    0.0,
 			expectedRejections: 10.0,
-			filterPolicies: []sharedconfig.FilterPolicy{
+			filterPolicies: []filterconfig.FilterPolicy{
 				{
 
-					Include: &sharedconfig.PolicyMatch{
-						MatchType: sharedconfig.Strict,
-						Attributes: []sharedconfig.MatchPolicyAttribute{
+					Include: &filterconfig.PolicyMatch{
+						MatchType: filterconfig.Strict,
+						Attributes: []filterconfig.MatchPolicyAttribute{
 							{
 								Key:   "span.nope",
 								Value: "nothere",
@@ -201,11 +200,11 @@ func TestSpanMetrics_applyFilterPolicy(t *testing.T) {
 		{
 			expectedMatches:    0.0,
 			expectedRejections: 10.0,
-			filterPolicies: []sharedconfig.FilterPolicy{
+			filterPolicies: []filterconfig.FilterPolicy{
 				{
-					Exclude: &sharedconfig.PolicyMatch{
-						MatchType: sharedconfig.Strict,
-						Attributes: []sharedconfig.MatchPolicyAttribute{
+					Exclude: &filterconfig.PolicyMatch{
+						MatchType: filterconfig.Strict,
+						Attributes: []filterconfig.MatchPolicyAttribute{
 							{
 								Key:   "status",
 								Value: "STATUS_CODE_OK",
@@ -218,11 +217,11 @@ func TestSpanMetrics_applyFilterPolicy(t *testing.T) {
 		{
 			expectedMatches:    10.0,
 			expectedRejections: 0.0,
-			filterPolicies: []sharedconfig.FilterPolicy{
+			filterPolicies: []filterconfig.FilterPolicy{
 				{
-					Include: &sharedconfig.PolicyMatch{
-						MatchType: sharedconfig.Regex,
-						Attributes: []sharedconfig.MatchPolicyAttribute{
+					Include: &filterconfig.PolicyMatch{
+						MatchType: filterconfig.Regex,
+						Attributes: []filterconfig.MatchPolicyAttribute{
 							{
 								Key:   "kind",
 								Value: "SPAN_KIND_.*",
@@ -293,441 +292,8 @@ func TestSpanMetrics_applyFilterPolicy(t *testing.T) {
 			assert.Equal(t, tc.expectedMatches, testRegistry.Query("traces_spanmetrics_latency_count", lbls))
 			assert.Equal(t, tc.expectedMatches, testRegistry.Query("traces_spanmetrics_latency_sum", lbls))
 			assert.Equal(t, tc.expectedRejections, testRegistry.Query(metricFilterDropsTotal, nil))
-
 		})
 	}
-
-}
-
-func TestSpanMetrics_policyMatch(t *testing.T) {
-	cases := []struct {
-		policy   *sharedconfig.PolicyMatch
-		resource *v1.Resource
-		span     *trace_v1.Span
-		expect   bool
-	}{
-		{
-			expect: true,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "span.kind",
-						Value: "client",
-					},
-					{
-						Key:   "resource.location",
-						Value: "earth",
-					},
-					{
-						Key:   "resource.name",
-						Value: "goodiegoodie",
-					},
-					{
-						Key:   "resource.othervalue",
-						Value: "somethinginteresting",
-					},
-				},
-			},
-			resource: &v1.Resource{
-				Attributes: []*common_v1.KeyValue{
-					{
-						Key: "name",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "goodiegoodie",
-							},
-						},
-					},
-					{
-						Key: "location",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "earth",
-							},
-						},
-					},
-					{
-						Key: "othervalue",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "somethinginteresting",
-							},
-						},
-					},
-				},
-			},
-			span: &trace_v1.Span{
-				Attributes: []*common_v1.KeyValue{
-					{
-						Key: "kind",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "client",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			expect: true,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "span.kind",
-						Value: "client",
-					},
-				},
-			},
-			resource: &v1.Resource{
-				Attributes: []*common_v1.KeyValue{
-					{
-						Key: "location",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "earth",
-							},
-						},
-					},
-					{
-						Key: "othervalue",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "somethinginteresting",
-							},
-						},
-					},
-				},
-			},
-			span: &trace_v1.Span{
-				Attributes: []*common_v1.KeyValue{
-					{
-						Key: "kind",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "client",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			expect: true,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "resource.location",
-						Value: "earth",
-					},
-					{
-						Key:   "resource.othervalue",
-						Value: "somethinginteresting",
-					},
-				},
-			},
-			resource: &v1.Resource{
-				Attributes: []*common_v1.KeyValue{
-					{
-						Key: "location",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "earth",
-							},
-						},
-					},
-					{
-						Key: "othervalue",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "somethinginteresting",
-							},
-						},
-					},
-				},
-			},
-			span: &trace_v1.Span{
-				Attributes: []*common_v1.KeyValue{
-					{
-						Key: "kind",
-						Value: &common_v1.AnyValue{
-							Value: &common_v1.AnyValue_StringValue{
-								StringValue: "client",
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tc := range cases {
-		r := policyMatch(getSplitPolicy(tc.policy), tc.resource, tc.span)
-		require.Equal(t, tc.expect, r)
-	}
-}
-
-func TestSpanMetrics_policyMatchIntrinsicAttrs(t *testing.T) {
-	cases := []struct {
-		policy *sharedconfig.PolicyMatch
-		span   *trace_v1.Span
-		expect bool
-	}{
-		{
-			expect: true,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "kind",
-						Value: "SPAN_KIND_SERVER",
-					},
-					{
-						Key:   "status",
-						Value: "STATUS_CODE_OK",
-					},
-					{
-						Key:   "name",
-						Value: "goodiegoodie",
-					},
-				},
-			},
-			span: &trace_v1.Span{
-				Kind: trace_v1.Span_SPAN_KIND_SERVER,
-				Status: &trace_v1.Status{
-					Code: trace_v1.Status_STATUS_CODE_OK,
-				},
-				Name: "goodiegoodie",
-			},
-		},
-	}
-
-	for _, tc := range cases {
-		r := policyMatchIntrinsicAttrs(tc.policy, tc.span)
-		require.Equal(t, tc.expect, r)
-	}
-
-}
-
-func TestSpanMetrics_policyMatchAttrs(t *testing.T) {
-	cases := []struct {
-		policy *sharedconfig.PolicyMatch
-		attrs  []*common_v1.KeyValue
-		expect bool
-	}{
-		// Single string match
-		{
-			expect: true,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "foo",
-						Value: "bar",
-					},
-				},
-			},
-			attrs: []*common_v1.KeyValue{
-				{
-					Key: "foo",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_StringValue{
-							StringValue: "bar",
-						},
-					},
-				},
-			},
-		},
-		// Multiple string match
-		{
-			expect: true,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "foo",
-						Value: "bar",
-					},
-					{
-						Key:   "otherfoo",
-						Value: "notbar",
-					},
-				},
-			},
-			attrs: []*common_v1.KeyValue{
-				{
-					Key: "foo",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_StringValue{
-							StringValue: "bar",
-						},
-					},
-				},
-				{
-					Key: "otherfoo",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_StringValue{
-							StringValue: "notbar",
-						},
-					},
-				},
-			},
-		},
-		// Multiple string non match
-		{
-			expect: false,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "foo",
-						Value: "bar",
-					},
-					{
-						Key:   "otherfoo",
-						Value: "nope",
-					},
-				},
-			},
-			attrs: []*common_v1.KeyValue{
-				{
-					Key: "foo",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_StringValue{
-							StringValue: "bar",
-						},
-					},
-				},
-				{
-					Key: "otherfoo",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_StringValue{
-							StringValue: "notbar",
-						},
-					},
-				},
-			},
-		},
-		// Combination match
-		{
-			expect: true,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "one",
-						Value: "1",
-					},
-					{
-						Key:   "oneone",
-						Value: 11,
-					},
-					{
-						Key:   "oneonepointone",
-						Value: 11.1,
-					},
-					{
-						Key:   "matching",
-						Value: true,
-					},
-				},
-			},
-			attrs: []*common_v1.KeyValue{
-				{
-					Key: "one",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_StringValue{
-							StringValue: "1",
-						},
-					},
-				},
-				{
-					Key: "oneone",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_IntValue{
-							IntValue: 11,
-						},
-					},
-				},
-				{
-					Key: "oneonepointone",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_DoubleValue{
-							DoubleValue: 11.1,
-						},
-					},
-				},
-				{
-					Key: "matching",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_BoolValue{
-							BoolValue: true,
-						},
-					},
-				},
-			},
-		},
-		// Regex basic match
-		{
-			expect: true,
-			policy: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Regex,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
-					{
-						Key:   "dd",
-						Value: `\d\d\w{5}`,
-					},
-				},
-			},
-			attrs: []*common_v1.KeyValue{
-				{
-					Key: "dd",
-					Value: &common_v1.AnyValue{
-						Value: &common_v1.AnyValue_StringValue{
-							StringValue: "11xxxxx",
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tc := range cases {
-		r := policyMatchAttrs(tc.policy, tc.attrs)
-		require.Equal(t, tc.expect, r)
-	}
-
-}
-
-func TestSpanMetrics_stringMatch(t *testing.T) {
-	cases := []struct {
-		matchType sharedconfig.MatchType
-		s         string
-		pattern   string
-		expect    bool
-	}{
-		{
-			matchType: sharedconfig.Strict,
-			s:         "foo",
-			pattern:   "foo",
-			expect:    true,
-		},
-		{
-			matchType: sharedconfig.Strict,
-			s:         "foo",
-			pattern:   "bar",
-			expect:    false,
-		},
-	}
-
-	for _, tc := range cases {
-		r := stringMatch(tc.matchType, tc.s, tc.pattern)
-		require.Equal(t, tc.expect, r)
-	}
-
 }
 
 func withLe(lbls labels.Labels, le float64) labels.Labels {
@@ -754,7 +320,7 @@ func BenchmarkSpanMetrics_applyFilterPolicyNone(b *testing.B) {
 	// b.Logf("size: %s", humanize.Bytes(uint64(batch.Size())))
 	// b.Logf("span count: %d", len(batch.ScopeSpans))
 
-	policies := []sharedconfig.FilterPolicy{}
+	policies := []filterconfig.FilterPolicy{}
 
 	benchmarkFilterPolicy(b, policies, batch)
 }
@@ -767,11 +333,11 @@ func BenchmarkSpanMetrics_applyFilterPolicySmall(b *testing.B) {
 	err = batch.Unmarshal(data)
 	require.NoError(b, err)
 
-	policies := []sharedconfig.FilterPolicy{
+	policies := []filterconfig.FilterPolicy{
 		{
-			Include: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
+			Include: &filterconfig.PolicyMatch{
+				MatchType: filterconfig.Strict,
+				Attributes: []filterconfig.MatchPolicyAttribute{
 					{
 						Key:   "span.foo",
 						Value: "foo-value",
@@ -792,11 +358,11 @@ func BenchmarkSpanMetrics_applyFilterPolicyMedium(b *testing.B) {
 	err = batch.Unmarshal(data)
 	require.NoError(b, err)
 
-	policies := []sharedconfig.FilterPolicy{
+	policies := []filterconfig.FilterPolicy{
 		{
-			Include: &sharedconfig.PolicyMatch{
-				MatchType: sharedconfig.Strict,
-				Attributes: []sharedconfig.MatchPolicyAttribute{
+			Include: &filterconfig.PolicyMatch{
+				MatchType: filterconfig.Strict,
+				Attributes: []filterconfig.MatchPolicyAttribute{
 					{
 						Key:   "span.foo",
 						Value: "foo-value",
@@ -821,7 +387,7 @@ func BenchmarkSpanMetrics_applyFilterPolicyMedium(b *testing.B) {
 	benchmarkFilterPolicy(b, policies, batch)
 }
 
-func benchmarkFilterPolicy(b *testing.B, policies []sharedconfig.FilterPolicy, batch *trace_v1.ResourceSpans) {
+func benchmarkFilterPolicy(b *testing.B, policies []filterconfig.FilterPolicy, batch *trace_v1.ResourceSpans) {
 	testRegistry := registry.NewTestRegistry()
 	cfg := Config{}
 	cfg.RegisterFlagsAndApplyDefaults("", nil)
